Re-locate theme theme context Relocate theme context to the context directory. Modify components and tests to reference themeContext
